Flyttar serialiseringsopplegget inn i helper-klassa
